added implementation for multiple delete

+++ src/com/tasma/commands/DeleteCommand.java
@@ -5,38 +5,67 @@
 package com.tasma.commands;
 
 import java.util.List;
-
+import java.util.LinkedList;
 import com.tasma.Palette;
 import com.tasma.Task;
 import com.tasma.TaskCollection;
 import com.tasma.TasmaUserInterface;
 import com.tasma.UIMessage;
 
+
 public class DeleteCommand extends AbstractUndoableCommand  {
 
 	private List<Task> state;
-	private Task task;
+	private List<Task> tasks;
 	
 	public DeleteCommand(TasmaUserInterface userInterface,
-			TaskCollection collection, List<Task> state, int index) {
+			TaskCollection collection, List<Task> state, List<Integer> indices) {
 		super(userInterface, collection);
 		this.state = state;
-		this.task = state.get(index);
+		tasks = new LinkedList<Task>();
+		for (int index: indices) {
+			tasks.add(state.get(index));
+		}
 	}
 
 	@Override
 	public void execute() throws Exception {
-		collection.delete(task);
-		userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_SUCCESS, task.getDetails()), Palette.MESSAGE_SUCCESS);
-		
-		ListCommand listCommand = new ListCommand(userInterface, collection, state);
-		listCommand.execute();
+		if (tasks.size() == 0) {
+			throw new NotExecutedException();
+		} else {
+			if (tasks.size() == 1) {
+				Task task = tasks.get(0);
+				collection.delete(task);
+
+				userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_SUCCESS, task.getDetails()), Palette.MESSAGE_SUCCESS);
+			} else {
+				for (Task task : tasks) {
+					collection.delete(task);
+				}
+				userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_MULTIPLE_SUCCESS, tasks.size()), Palette.MESSAGE_SUCCESS);
+			}
+			
+			ListCommand listCommand = new ListCommand(userInterface, collection, state);
+			listCommand.execute();
+		}
 	}
 	
 	@Override
 	public void undo() throws Exception {
-		collection.create(task);
-
-		userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_UNDO, task.getDetails()), Palette.MESSAGE_SUCCESS);
+		if (tasks.size() == 0) {
+			throw new NotExecutedException();
+		} else {
+			if (tasks.size() == 1) {
+				Task task = tasks.get(0);
+				collection.create(task);
+		
+				userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_UNDO, task.getDetails()), Palette.MESSAGE_SUCCESS);
+			} else {
+				for (Task task : tasks) {
+					collection.create(task);
+				}
+				userInterface.displayMessage(String.format(UIMessage.COMMAND_DELETE_MULTIPLE_UNDO, tasks.size()), Palette.MESSAGE_SUCCESS);
+			}
+		}
 	}
 }
